Fixed crash on schedule caching preference first-time enable
Really odd mistake on my part when I originally wrote the schedule database:
createDatabase() was catching an IOException and throwing back an Error. Not
really sure why I thought to do that, but it was triggering a crash and only
seemed to happen on the first time the schedule caching preference was enabled.