Correctly handle loading multiple class listings. Prevents buggy behavior such as loading multiple copies of the same class listing from happening.
Honestly, this is still being done terribly. I definitely should not be
accessing a parent activity's state from a fragment in such a clunky way.
I should be parsing part of the page in the parent activity, and then
spawning more fragments from there as necessary.  But hey, it works :)