Schedule: Implement correct persistence in CourseScheduleFragment
Big helpers were Otto and ArrayPagerAdapter.
Also did a general cleanup of code (even some minor stuff in
ExamScheduleFragment snuck in)
This required a surprisingly massive rework of data mocking, but I think it's
more robust/general now because of it.
Currently auto-re-logging in is disabled, may not re-enable it, we'll see.