Fix Data Usage crash in Android <4.2
I believe this issue was due to Otto enumerating all
annotated methods in a class. When it hit a method
annotated with @JavascriptInterface on an Android version
that does not yet support that class, it would crash with
NoClassDefError. Putting the annotated method into its own
class has fixed things.